# swagger.yml
openapi: 3.0.2
info:
    title: movies_api
    version: 1.0.0
paths:
    /movies:
        summary: Path used to manage the list of movies.
        description: 'The REST endpoint/path used to list and create zero or more `Movie` entities. '
        get:
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Movie'
                            examples:
                                200_OK:
                                    value:
                                        -
                                            id: 1
                                            name: movie1
                                            rating: 3
                                            year: 2004
                                            poster: /movie-data/posters/1.png
                                        -
                                            id: 2
                                            name: movie2
                                            rating: 4
                                            year: 1993
                                            poster: /movie-data/posters/2.png
                                        -
                                            id: 3
                                            name: movie3
                                            rating: 5
                                            year: 2022
                                            poster: /movie-data/posters/3.png
                    description: Successful response - returns an array of `Movie` entities.
            operationId: getMovies
            summary: Get a list of every movie
            description: Gets a list of all `Movie` entities.
        post:
            requestBody:
                description: A new Movie to be created.
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Movie'
                        examples:
                            '204':
                                value:
                                    id: 1
                                    name: movie1
                                    rating: 3
                                    year: 2004
                                    poster: >-
                                        data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAgAAAAIAQMAAAD+wSzIAAAABlBMVEX///+/v7+jQ3Y5AAAADklEQVQI12P4AIX8EAgALgAD/aNpbtEAAAAASUVORK5CYII
                            '400':
                                value:
                                    id: 1
                                    name: ''
                                    rating: 3
                                    year: 2004
                                    poster: >-
                                        data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAgAAAAIAQMAAAD+wSzIAAAABlBMVEX///+/v7+jQ3Y5AAAADklEQVQI12P4AIX8EAgALgAD/aNpbtEAAAAASUVORK5CYII
                required: true
            responses:
                '201':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Movie'
                            examples:
                                '201':
                                    value: ''
                    description: Successful response - Adds movie to the list of movies
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Movie'
                            examples:
                                400_BAD_JSON:
                                    value:
                                        error: expected end of JSON input
                                400_MISSING_PARAMS:
                                    value:
                                        error: One or all params are missing
                                400_MISSING_TITLE:
                                    value:
                                        error: Missing Title
                    description: When the user is admin but one or more params was incorrect or missing
                '401':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Movie'
                            examples:
                                401_UNAUTHORIZED:
                                    value:
                                        error: Unauthorized
                                401_MISSING_AUTH_DATA:
                                    value:
                                        error: Missing authorization data
                    description: Will not add a movie to the list of movies because the user isn't logged in
                '403':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Movie'
                            examples:
                                403_FORBIDDEN:
                                    value:
                                        error: Forbidden
                    description: Will not add a movie to the list of movies because the user isnt an admin
            security:
                -
                    apiKey: []
            operationId: addMovie
            summary: Add a movie to the list
        delete:
            requestBody:
                content:
                    application/json:
                        schema:
                            type: array
                            items: {}
                        examples:
                            '204':
                                value:
                                    id: 1
                required: true
            responses:
                '204':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                '204':
                                    value: ''
                    description: Successful response - removes the selected movie from the list
                '400':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                400_BAD_JSON:
                                    value:
                                        error: Unexpected end of JSON input
                    description: When the request body JSON is incorrect.
                '401':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                401_UNAUTHORIZED:
                                    value:
                                        error: Unauthorized
                                401_MISSING_AUTH_DATA:
                                    value:
                                        error: Missing authorization data
                    description: Will not remove a movie from the list of movies because the user isn't logged in
                '403':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                403_FORBIDDEN:
                                    value:
                                        error: Forbidden
                    description: Will not remove a movie from the list of movies because the user isnt an admin
            security:
                -
                    apiKey: []
            operationId: deleteMovie
            summary: Remove a movie from the list
        patch:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Movie'
                        examples:
                            patch_204:
                                value:
                                    id: 1
                                    name: movie1
                                    rating: 3
                                    year: 2004
                                    poster: >-
                                        data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAgAAAAIAQMAAAD+wSzIAAAABlBMVEX///+/v7+jQ3Y5AAAADklEQVQI12P4AIX8EAgALgAD/aNpbtEAAAAASUVORK5CYII
                            patch_400:
                                value:
                                    id: 1
                                    name: ''
                                    rating: 3
                                    year: 2004
                                    poster: >-
                                        data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAgAAAAIAQMAAAD+wSzIAAAABlBMVEX///+/v7+jQ3Y5AAAADklEQVQI12P4AIX8EAgALgAD/aNpbtEAAAAASUVORK5CYII
                required: true
            responses:
                '204':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                '204':
                                    value: ''
                    description: Successful response - Modifies a movie in the list of movies
                '400':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                400_BAD_JSON:
                                    value:
                                        error: Unexpected end of JSON input
                                400_INVALID_TITLE:
                                    value:
                                        error: Invalid title
                    description: >-
                        When the user is admin and title WAS given but was EMPTY OR the request body was
                        invalid JSON
                '401':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                401_UNAUTHORIZED:
                                    value:
                                        error: Unauthorized
                                401_MISSING_AUTH_DATA:
                                    value:
                                        error: Missing authorization data
                    description: When the user was not logged in
                '403':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                403_FORBIDDEN:
                                    value:
                                        error: Forbidden
                    description: When the user was not admin
            security:
                -
                    apiKey: []
            operationId: editMovie
            summary: Edit movie info in the list
    /logs:
        get:
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                200_OK:
                                    value:
                                        -
                                            log_id: '1'
                                            timestamp: '2023-04-26 20:10:08'
                                            http_method: DELETE
                                        -
                                            log_id: '2'
                                            timestamp: '2023-04-27 13:44:27'
                                            http_method: DELETE
                    description: Successful response - returns an array of `Logs` entities.
            operationId: getLogs
            summary: Gets a list of logs affecting movies
            description: Gets a list of all `Logs` entities.
    /sessions:
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            type: array
                            items:
                                type: string
                        examples:
                            200_OK:
                                value:
                                    username: User
                                    password: Password
                            400_MISSING_USERNAME:
                                value:
                                    password: Password
                            400_MISSING_PASSWORD:
                                value:
                                    username: User
                            401_INCORRECT_PASSWORD:
                                value:
                                    username: User
                                    password: wrong
                            401_UNREGISTERED:
                                value:
                                    username: unregistered
                                    password: Password
                required: true
            responses:
                '201':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                Successful login:
                                    value:
                                        sessionId: 1
                                        isAdmin: true
                    description: Successful response - signs the user in.
                '400':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                Missing login credentials:
                                    value:
                                        error: Missing login credentials
                                Login unsuccessful:
                                    value:
                                        error: Login unsuccessful
                    description: When authentication data was missing or something went wrong.
                '401':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                'Unauthorized: The password you entered is incorrect':
                                    value:
                                        error: 'Unauthorized: The password you entered is incorrect'
                                'Unauthorized: The email you entered is not registered':
                                    value:
                                        error: 'Unauthorized: The email you entered is not registered'
                    description: When the authentication data was provided but the details weren't correct.
            deprecated: false
            operationId: signIn
            summary: Signs the user in.
        delete:
            requestBody:
                required: true
            responses:
                '204':
                    description: Successful response - user is logged out.
            operationId: signOut
            summary: Signs the user out.
    /env:
        get:
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                200_OK:
                                    value:
                                        client_id: 123ABC
                    description: Successful response - returns an object of .env values.
            summary: Gets .env values for the client-side to use.
components:
    schemas:
        Movie:
            title: Root Type for Movie
            description: ''
            type: object
            properties:
                id:
                    format: int32
                    type: integer
                name:
                    type: string
                rating:
                    format: int32
                    type: integer
                year:
                    format: int32
                    type: integer
                poster:
                    type: string
            example:
                id: 1
                name: movie1
                rating: 3
                year: 2004
                poster: >-
                    data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAgAAAAIAQMAAAD+wSzIAAAABlBMVEX///+/v7+jQ3Y5AAAADklEQVQI12P4AIX8EAgALgAD/aNpbtEAAAAASUVORK5CYII
    securitySchemes:
        Bearer:
            scheme: bearer
            type: http
        apiKey:
            type: apiKey
            name: Authorization
            in: header